+++ CHANGELOG.md
@@ -5,6 +5,13 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+* `HandshakeError::IncompatibleProtocol` carries both the current version and
+  the required version.
+
 ## [0.2.7] - 2022-06-08
 
 ### Added
@@ -77,6 +84,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 * `send_frame` and `recv_frame` to send and receive length-delimited frames.
 
+[Unreleased]: https://github.com/petabi/oinq/compare/0.2.7...main
 [0.2.7]: https://github.com/petabi/oinq/compare/0.2.6...0.2.7
 [0.2.6]: https://github.com/petabi/oinq/compare/0.2.5...0.2.6
 [0.2.5]: https://github.com/petabi/oinq/compare/0.2.4...0.2.5